GMT+1 bug fix
The .ToUniversalTime() resulted in events being created a day too soon.
EventDate was at midnight, a call to ToUniversalTime took an hour off
now I'm in BST, a day off that results in it being two days before the
collection. I've now changed this to enforce it being defined as a Utc
time.